Fix compiling for ARM64

+++ src/libs/aroma_graph.c
@@ -1104,7 +1104,7 @@ void ag_sync_force() {
   }
 }
 static void * ag_sync_fade_thread(void * cookie) {
-  int frame = (int) cookie;
+  long frame = (long) cookie;
   ag_isbusy = 0;
   ag_sync_locked = 1;
   ag_refreshlock = 1;
@@ -1177,11 +1177,11 @@ static void * ag_sync_fade_thread(void * cookie) {
   return NULL;
 }
 void ag_sync_fade_wait(int frame) {
-  ag_sync_fade_thread((void *) frame);
+  ag_sync_fade_thread((void *)(uintptr_t) frame);
 }
 void ag_sync_fade(int frame) {
   pthread_t threadsyncfade;
-  pthread_create(&threadsyncfade, NULL, ag_sync_fade_thread, (void *) frame);
+  pthread_create(&threadsyncfade, NULL, ag_sync_fade_thread, (void *)(uintptr_t) frame);
   pthread_detach(threadsyncfade);
 }
 byte ag_blur_h(CANVAS * d, CANVAS * s, int radius) {
